moved the computer player class TicTacToePlaye into players.py and renamed it to ComputerPlayerO
